+++ consensus/pacemaker.go
@@ -2,12 +2,12 @@ package consensus
 
 import (
 	"context"
-	"fmt"
 	"log"
 	timePkg "time"
 
 	consensusTelemetry "github.com/pokt-network/pocket/consensus/telemetry"
 	typesCons "github.com/pokt-network/pocket/consensus/types"
+	"github.com/pokt-network/pocket/runtime/configs"
 	"github.com/pokt-network/pocket/shared/modules"
 )
 
@@ -31,10 +31,8 @@ type Pacemaker interface {
 }
 
 var (
-	_ modules.Module             = &paceMaker{}
-	_ modules.ConfigurableModule = &paceMaker{}
-	_ PacemakerDebug             = &paceMaker{}
-	_ modules.PacemakerConfig    = &typesCons.PacemakerConfig{}
+	_ modules.Module = &paceMaker{}
+	_ PacemakerDebug = &paceMaker{}
 )
 
 type paceMaker struct {
@@ -46,7 +44,7 @@ type paceMaker struct {
 	// a great idea in production code.
 	consensusMod *consensusModule
 
-	pacemakerCfg modules.PacemakerConfig
+	pacemakerCfg *configs.PacemakerConfig
 
 	stepCancelFunc context.CancelFunc
 
@@ -61,11 +59,8 @@ func CreatePacemaker(runtimeMgr modules.RuntimeMgr) (modules.Module, error) {
 
 func (m *paceMaker) Create(runtimeMgr modules.RuntimeMgr) (modules.Module, error) {
 	cfg := runtimeMgr.GetConfig()
-	if err := m.ValidateConfig(cfg); err != nil {
-		log.Fatalf("config validation failed: %v", err)
-	}
 
-	pacemakerCfg := cfg.GetConsensusConfig().(HasPacemakerConfig).GetPacemakerConfig()
+	pacemakerCfg := cfg.Consensus.PacemakerConfig
 
 	return &paceMaker{
 		bus:          nil,
@@ -76,8 +71,8 @@ func (m *paceMaker) Create(runtimeMgr modules.RuntimeMgr) (modules.Module, error
 		stepCancelFunc: nil, // Only set on restarts
 
 		paceMakerDebug: paceMakerDebug{
-			manualMode:                pacemakerCfg.GetManual(),
-			debugTimeBetweenStepsMsec: pacemakerCfg.GetDebugTimeBetweenStepsMsec(),
+			manualMode:                pacemakerCfg.Manual,
+			debugTimeBetweenStepsMsec: pacemakerCfg.DebugTimeBetweenStepsMsec,
 			quorumCertificate:         nil,
 		},
 	}, nil
@@ -104,13 +99,6 @@ func (m *paceMaker) GetBus() modules.Bus {
 		log.Fatalf("PocketBus is not initialized")
 	}
 	return m.bus
-}
-
-func (*paceMaker) ValidateConfig(cfg modules.Config) error {
-	if _, ok := cfg.GetConsensusConfig().(HasPacemakerConfig); !ok {
-		return fmt.Errorf("cannot cast to PacemakeredConsensus")
-	}
-	return nil
 }
 
 func (m *paceMaker) SetConsensusModule(c *consensusModule) {
@@ -259,6 +247,6 @@ func (p *paceMaker) startNextView(qc *typesCons.QuorumCertificate, forceNextView
 
 // TODO(olshansky): Increase timeout using exponential backoff.
 func (p *paceMaker) getStepTimeout(round uint64) timePkg.Duration {
-	baseTimeout := timePkg.Duration(int64(timePkg.Millisecond) * int64(p.pacemakerCfg.GetTimeoutMsec()))
+	baseTimeout := timePkg.Duration(int64(timePkg.Millisecond) * int64(p.pacemakerCfg.TimeoutMsec))
 	return baseTimeout
 }