+++ persistence/module.go
@@ -8,7 +8,8 @@ import (
 	"github.com/jackc/pgx/v5"
 	"github.com/pokt-network/pocket/persistence/indexer"
 	"github.com/pokt-network/pocket/persistence/kvstore"
-	"github.com/pokt-network/pocket/persistence/types"
+	"github.com/pokt-network/pocket/runtime/configs"
+	"github.com/pokt-network/pocket/runtime/genesis"
 	"github.com/pokt-network/pocket/shared/modules"
 )
 
@@ -16,17 +17,15 @@ var (
 	_ modules.PersistenceModule = &persistenceModule{}
 	_ modules.PersistenceModule = &persistenceModule{}
 
-	_ modules.PersistenceRWContext    = &PostgresContext{}
-	_ modules.PersistenceGenesisState = &types.PersistenceGenesisState{}
-	_ modules.PersistenceConfig       = &types.PersistenceConfig{}
+	_ modules.PersistenceRWContext = &PostgresContext{}
 )
 
 // TODO: convert address and public key to string not bytes in all account and actor functions
 // TODO: remove address parameter from all pool operations
 type persistenceModule struct {
 	bus          modules.Bus
-	config       modules.PersistenceConfig
-	genesisState modules.PersistenceGenesisState
+	config       *configs.PersistenceConfig
+	genesisState *genesis.GenesisState
 
 	blockStore kvstore.KVStore
 	txIndexer  indexer.TxIndexer
@@ -47,7 +46,8 @@ func Create(runtimeMgr modules.RuntimeMgr) (modules.Module, error) {
 func (*persistenceModule) Create(runtimeMgr modules.RuntimeMgr) (modules.Module, error) {
 	var m *persistenceModule
 
-	cfg := runtimeMgr.GetConfig()
+	persistenceCfg := runtimeMgr.GetConfig().Persistence
+	genesisState := runtimeMgr.GetGenesis()
 
 	if err := m.ValidateConfig(cfg); err != nil {
 		return nil, fmt.Errorf("config validation failed: %w", err)
@@ -70,17 +70,17 @@ func (*persistenceModule) Create(runtimeMgr modules.RuntimeMgr) (modules.Module,
 	conn.Close(context.TODO())
 
 	// TODO: Follow the same pattern as txIndexer below for initializing the blockStore
-	blockStore, err := initializeBlockStore(persistenceCfg.GetBlockStorePath())
+	blockStore, err := initializeBlockStore(persistenceCfg.BlockStorePath)
 	if err != nil {
 		return nil, err
 	}
 
-	txIndexer, err := indexer.NewTxIndexer(persistenceCfg.GetTxIndexerPath())
+	txIndexer, err := indexer.NewTxIndexer(persistenceCfg.TxIndexerPath)
 	if err != nil {
 		return nil, err
 	}
 
-	stateTrees, err := newStateTrees(persistenceCfg.GetTreesStoreDir())
+	stateTrees, err := newStateTrees(persistenceCfg.TreesStoreDir)
 	if err != nil {
 		return nil, err
 	}
@@ -88,7 +88,7 @@ func (*persistenceModule) Create(runtimeMgr modules.RuntimeMgr) (modules.Module,
 	m = &persistenceModule{
 		bus:          nil,
 		config:       persistenceCfg,
-		genesisState: persistenceGenesis,
+		genesisState: genesisState,
 
 		blockStore: blockStore,
 		txIndexer:  txIndexer,
@@ -104,7 +104,7 @@ func (*persistenceModule) Create(runtimeMgr modules.RuntimeMgr) (modules.Module,
 	if shouldHydrateGenesis, err := m.shouldHydrateGenesisDb(); err != nil {
 		return nil, err
 	} else if shouldHydrateGenesis {
-		m.populateGenesisState(persistenceGenesis) // fatal if there's an error
+		m.populateGenesisState(genesisState) // fatal if there's an error
 	} else {
 		// This configurations will connect to the SQL database and key-value stores specified
 		// in the configurations and connected to those.
@@ -137,16 +137,6 @@ func (m *persistenceModule) GetBus() modules.Bus {
 		log.Fatalf("PocketBus is not initialized")
 	}
 	return m.bus
-}
-
-func (*persistenceModule) ValidateConfig(cfg modules.Config) error {
-	// TODO (#334): implement this
-	return nil
-}
-
-func (*persistenceModule) ValidateGenesis(genesis modules.GenesisState) error {
-	// TODO (#334): implement this
-	return nil
 }
 
 func (m *persistenceModule) NewRWContext(height int64) (modules.PersistenceRWContext, error) {