+++ consensus/doc/CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.15] - 2022-12-28
+
+- ValidatorMap uses `Actor` references now
+
+## [0.0.0.14] - 2022-12-21
+
+- Updated do use the new centralized config and genesis
+- `Actor` is now a shared `struct` instead of an `interface`
+- Removed converters between the interfaces and the consensus structs for Validators
+
 ## [0.0.0.13] - 2022-12-14
 
 - Consolidated number of validators in tests in a single constant: `numValidators`