+++ utility/doc/CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.15] - 2022-12-28
+
+- Renamed enum names as per code-review
+- Using defaults from `test_artifacts` for tests
+- Updated tests to reflect the above changes
+
+## [0.0.0.14] - 2022-12-21
+
+- Updated to use the new centralized config and genesis handling
+- Updated to use the new `Actor` struct under `coreTypes`
+- Updated tests and mocks
+
 ## [0.0.0.13] - 2022-12-10
 
 - Introduce `SetProposalBlock` and local vars to keep proposal state